requirements.txt
python-telegram-bot==21.5
requests
matplotlib
pandas
numpy